// src/definitions.d.ts
export type AsyncHandler<T, TResult> = (value : T, wasResolved : boolean) => TResult | PromiseLike<TResult>;
// ...
export interface ExtendedPromiseConstructor {

    /**
     * A reference to the prototype.
     */
    readonly prototype: Promise<any>;

    /**
     * Creates a new Promise.
     * @param executor A callback used to initialize the promise. This callback is passed two arguments:
     * a resolve callback used resolve the promise with a value or the result of another promise,
     * and a reject callback used to reject the promise with a provided reason or error.
     */
    new <T>(executor: (resolve: (value?: T | PromiseLike<T>) => void, reject: (reason?: any) => void) => void): ExtendedPromise<T>;

    /**
     * Schedules an action to occur immediately but after current code has finished executing, using the Promise's scheduling mechanism.
     * This is similar to `setTimeout(action, 0)` or `setImmediate(action)`.
     * @template T
     * @param {() => (T | Promise<T>)} action The acion to execute.
     * @returns {Promise<T>}
     */
    soon<T>(action : () => (T | PromiseLike<T>)) : ExtendedPromise<T>;

    /**
     * Converts the given Promise/A+ implementation into ExtendedPromise promise using this constructor.
     * @param {PromiseLike<T>} other The Promise/A+ implementation.
     * @returns {ExtendedPromise<T>}
     */
    from<T>(other : PromiseLike<T>) : ExtendedPromise<T>

    /**
     * Returns a promise that will never resolve or reject.
     * @returns {PromiseLike<never>}
     */
    never() : ExtendedPromise<never>;

    /**
     * Returns a promise that will wait for some time, and then resolve with the value given in `value` (undefined by default).
     * @param {number | Date} deadline If a number, the milliseconds to wait. If a date, the date to wait for.
     * @param {T} value The value to resolve with.
     * @template T
     * @returns {PromiseLike<T>}
     */
    wait<T = void>(deadline : number | Date, value ?: T) : ExtendedPromise<T>;

    /**
     * Constructs a new promise, similarly to the Promise constructor. The main difference is that `definition` itself may return a promise (or be async).
     * If the promise returned by `definition` rejects, the returned promise will also reject. Normally, this would cause the returned promise to hang.
     * @template T
     * @param {(resolve: (x: T) => void, reject: (y: T) => void) => (void | PromiseLike<void>)} executor
     * @returns {PromiseLike<T>}
     */
    create<T>(executor : (resolve : ((x : T) => void), reject : ((y : T) => void)) => void | Promise<void>) : ExtendedPromise<T>;
}